+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Connector/Tasklet/EvaluationJobAwareTrait.php
@@ -1,0 +1,33 @@
+<?php
+
+/*
+ * This file is part of the Akeneo PIM Enterprise Edition.
+ *
+ * (c) 2020 Akeneo SAS (http://www.akeneo.com)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Connector\Tasklet;
+
+use Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Connector\JobParameters\PrepareEvaluationsParameters;
+use Akeneo\Tool\Component\Batch\Model\StepExecution;
+
+trait EvaluationJobAwareTrait
+{
+    /** @var StepExecution */
+    private $stepExecution;
+
+    public function setStepExecution(StepExecution $stepExecution)
+    {
+        $this->stepExecution = $stepExecution;
+    }
+
+    private function updatedSince(): \DateTimeImmutable
+    {
+        $updatedSince = $this->stepExecution->getJobParameters()->get(PrepareEvaluationsParameters::UPDATED_SINCE_PARAMETER);
+
+        return \DateTimeImmutable::createFromFormat(PrepareEvaluationsParameters::UPDATED_SINCE_DATE_FORMAT, $updatedSince);
+    }
+}