+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Connector/Tasklet/PimEnterpriseEvaluateProductModelsCriteriaTasklet.php
@@ -18,13 +18,14 @@ use Akeneo\Pim\Automation\DataQualityInsights\Application\ProductEvaluation\Crea
 use Akeneo\Pim\Automation\DataQualityInsights\Application\ProductEvaluation\EvaluatePendingCriteria;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\ProductEvaluation\GetProductIdsToEvaluateQueryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Repository\CriterionEvaluationRepositoryInterface;
-use Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Connector\JobParameters\PrepareEvaluationsParameters;
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
 use Akeneo\Tool\Component\Connector\Step\TaskletInterface;
 use Psr\Log\LoggerInterface;
 
 class PimEnterpriseEvaluateProductModelsCriteriaTasklet implements TaskletInterface
 {
+    use EvaluationJobAwareTrait;
+
     private const NB_PRODUCT_MODELS_MAX = 10000;
     private const BULK_SIZE = 100;
 
@@ -79,11 +80,6 @@ class PimEnterpriseEvaluateProductModelsCriteriaTasklet implements TaskletInterf
         }
     }
 
-    public function setStepExecution(StepExecution $stepExecution)
-    {
-        $this->stepExecution = $stepExecution;
-    }
-
     private function createMissingCriteriaEvaluations(): void
     {
         try {
@@ -99,13 +95,6 @@ class PimEnterpriseEvaluateProductModelsCriteriaTasklet implements TaskletInterf
                 ]
             );
         }
-    }
-
-    private function updatedSince(): \DateTimeImmutable
-    {
-        $evaluateFrom = $this->stepExecution->getJobParameters()->get(PrepareEvaluationsParameters::UPDATED_SINCE_PARAMETER);
-
-        return \DateTimeImmutable::createFromFormat(PrepareEvaluationsParameters::UPDATED_SINCE_DATE_FORMAT, $evaluateFrom);
     }
 
     private function cleanCriteriaOfDeletedProductModels()