+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - PIM-9392: Prevent generating asset thumbnail when file is too large
 - PIM-9411: Fix TWA project widget searching on all contributors
 - PIM-9372: Fix media-link thumbnail re-generation
+- PIM-9444: Fix locking issue on attribute table for retrieving the attribute options that need to be evaluated
 
 ## Improvements